+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- In rare circumstances outside of Recyclarr, quality profiles become invalid due to missing
+  required qualities. When this happens, users are not even able to save the profile using the
+  Sonarr or Radarr UI. Recyclarr now detects this situation and automatically repairs the quality
+  profile by re-adding these missing qualities for users. See [this issue][9738].
+
+[9738]: https://github.com/Radarr/Radarr/issues/9738
+
 ## [7.0.0] - 2024-06-27
 
 This release contains **BREAKING CHANGES**. See the [v7.0 Upgrade Guide][breaking7] for required